update MockServerSpec tests
